move athcon-client integration tests to its crate
