+++ ffi/athcon/bindings/rust/athcon-client/Cargo.toml
@@ -8,7 +8,12 @@ license.workspace = true
 edition.workspace = true
 
 [dependencies]
+athena-vmlib = { path = "../../../../vmlib" }
 athcon-sys = { path = "../athcon-sys" }
 athcon-vm = { path = "../athcon-vm" }
 enum_primitive = "0.1.1"
 num = "0.4"
+
+[dev-dependencies]
+hex = "0.4.3"
+athena-interface = { path = "../../../../../interface" }
